Fix for update from 9.6 to 9.6

+++ images/jobs/clean-postgres-data.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
+CURRENT_VERSION=$(cat /stolon-data/postgres/PG_VERSION)
+if [ $CURRENT_VERSION == $PG_VERSION_NEW ]; then exit 0; fi
+
 rm -rf /stolon-data/postgres
